requirements.txt
asgiref==3.11.0
Django==6.0
django-environ==0.12.0
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
